requirements.txt
# Core dependencies
python-dotenv
requests
beautifulsoup4
validators

# LangChain & OpenAI
langchain-core
langchain-openai
openai
langchain-community
langchain-text-splitters

# Vector Stores & Embeddings
faiss-cpu
langchain-chroma
langchain-huggingface
sentence_transformers

# Optional Enhancements
tqdm
aiohttp
numpy
pandas
torch
transformers
chromadb
duckduckgo-search

#Frontend
flask